Refactor get exif orientation to util
 - Move IPhoto to gallery types from event types
 - Use promisify and async await with exif-js
